+++ CHANGELOG.md
@@ -23,6 +23,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - [SIL.Windows.Forms] `CssLinkHref` property to `ShowReleaseNotesDialog` to allow linking to CSS file for
     displaying Markdown output.
 - [SIL.Scripture] IScrVerseRef interface (largely extracted from VerseRef)
+- [SIL.Core] IEnumerable<T>.FirstOrDefault extension method overload that takes a different default to use
+- [SIL.Core] string.IndexOf extension method overload that takes an IComparer<string>
 
 ### Changed